backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
websockets==13.1

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pgvector==0.3.2

# Redis / Task queue
redis==5.1.0
arq==0.26.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
qrcode[pil]==7.4.2
cryptography==43.0.3
python-magic==0.4.27

# AI Providers
openai==1.51.0
anthropic==0.34.2
google-generativeai==0.8.2

# Document processing
pypdf2==3.0.1
python-docx==1.1.2
pillow==10.4.0
pytesseract==0.3.13
pandas==2.2.3
openpyxl==3.1.5

# HTTP client
httpx==0.27.2
aiohttp==3.10.8

# Email
aiosmtplib==3.0.1
jinja2==3.1.4

# Validation & serialisation
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Utilities
python-dotenv==1.0.1
structlog==24.4.0
slowapi==0.1.9
sentry-sdk[fastapi]==2.14.0
tenacity==9.0.0
beautifulsoup4==4.12.3
lxml==5.3.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
pytest-cov==5.0.0
factory-boy==3.3.1
